backend/requirements.txt
# Dépendances nécessaires pour la détection (YOLOv8 + Detectron2)
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
numpy==1.26.4
Pillow==10.4.0
timm==0.9.12
ultralytics==8.3.36
opencv-python-headless==4.10.0.84
# Installation officielle Detectron2 CPU
# Voir https://detectron2.readthedocs.io/en/latest/tutorials/install.html
# Pour les systèmes Linux et macOS récents, la commande suivante fonctionne :
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
